README/OWNERS: update release creation section (#39)
* README: update release creation section

Updates the README file section for how to manually create a release.
Also adds a note that there's a periodic job that executes the release
creation.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* README: remove sentence about making PR ready for review

The GitHub action takes care of the PR and makes it ready for review, so
you don't have to any more. Therefore we delete that sentence.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>